set up config middleware
